requirements.txt
langchain
openai
chromadb
fastapi
uvicorn
python-dotenv
